refactor(document-service): Moving responses to their own package
Moving responses to their own package as there was starting to be too many to handle in the models section
