+++ packages/services/document-service/src/main/java/software/uncharted/terarium/documentserver/resources/xdd/DocumentResource.java
@@ -3,8 +3,9 @@ package software.uncharted.terarium.documentserver.resources.xdd;
 
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
-import software.uncharted.terarium.documentserver.models.xdd.XDDArticlesResponseOK;
-import software.uncharted.terarium.documentserver.models.xdd.XDDResponse;
+import software.uncharted.terarium.documentserver.responses.xdd.XDDArticlesResponseOK;
+import software.uncharted.terarium.documentserver.responses.xdd.XDDFacetsItemResponse;
+import software.uncharted.terarium.documentserver.responses.xdd.XDDResponse;
 import software.uncharted.terarium.documentserver.proxies.xdd.DocumentProxy;
 
 import javax.ws.rs.*;
@@ -70,9 +71,11 @@ public class DocumentResource {
 				match = null;
 			}
 
-			return proxy.getDocuments(
+			XDDResponse<XDDArticlesResponseOK> doc = proxy.getDocuments(
 				docid, doi, title, term, dataset, include_score, include_highlights, inclusive, full_results, max, per_page, dict, facets,
 				min_published, max_published, pubname, publisher, additional_fields, match);
+
+			return doc;
 
 		}
 		return new XDDResponse<>();
