+++ packages/services/document-service/src/main/java/software/uncharted/terarium/documentserver/responses/xdd/XDDRelatedDocumentsResponse.java
@@ -1,7 +1,8 @@
-package software.uncharted.terarium.documentserver.models.xdd;
+package software.uncharted.terarium.documentserver.responses.xdd;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
+import software.uncharted.terarium.documentserver.models.xdd.RelatedDocument;
 
 import java.util.List;
