+++ packages/services/document-service/src/main/java/software/uncharted/terarium/documentserver/responses/xdd/XDDSetsResponse.java
@@ -1,4 +1,4 @@
-package software.uncharted.terarium.documentserver.models.xdd;
+package software.uncharted.terarium.documentserver.responses.xdd;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
